+++ caffe_classifier/detect.cpp
@@ -183,7 +183,7 @@ void NNDetect<MatT>::generateInitialWindows(
 		float ball_real_size = 247.6; // ball is 9.75in diameter = 247.6 mm
         float percent_image = (float)wsize / scaledImages[scale].first.cols;
 		float size_fov = percent_image * hfov_; //TODO fov size
-		float depth_avg = (ball_real_size / (2.0 * tan(size_fov / 2.0))) - (4.572 * 25.4);
+		float depth_avg = (ball_real_size / (2.0 * tanf(size_fov / 2.0))) - (4.572 * 25.4);
 		
         float depth_min = depth_avg - depth_avg * depth_multiplier;
         float depth_max = depth_avg + depth_avg * depth_multiplier;
